modules/location-api/location-consumer/requirements.txt
SQLAlchemy==1.3.19
kafka-python
psycopg2-binary==2.9.5